data-engineering-projects/room_occupancy_detection_g5-main/load_service/requirements.txt
pymongo==4.3.3
kafka-python==2.0.2
